backend/requirements.txt
flask==3.0.2
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pymongo==4.6.2
bcrypt==4.1.2
weasyprint==61.2
Jinja2==3.1.3
python-dotenv==1.0.1